statement category service

+++ pennywise-model/src/main/resources/db.sql
@@ -1,6 +1,6 @@
-create table StatementCategory(
-	id integer primary key
-	,name varchar(40)
-);
-
-create unique index IDX_STATEMENT_CATEGORY_NAME on statementcategory(NAME);
+	create table StatementCategory(
+		id bigserial primary key
+		,name varchar(40)
+	);
+	
+	create unique index IDX_STATEMENT_CATEGORY_NAME on statementcategory(NAME);
